ai-service/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database dependencies
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# Basic text processing (lightweight)
textblob==0.17.1

# HTTP client
httpx==0.25.2

# Google Cloud (optional - will be handled gracefully if not available)
google-cloud-language==2.11.1
google-auth==2.23.4

# Scientific libraries (lighter versions)
numpy==1.24.4
scikit-learn==1.3.2 